Admin staff back-end

+++ modules/tables.php
@@ -1,7 +1,8 @@
 <?php
-    function formTableRows($selectQuary, $addChangeBtn, $addDeleteBtn){
+    //[ 'column' => 'join query' ]
+    function formTableRows($selectQuery, $addChangeBtn, $addDeleteBtn, $replaceJoinsCols = null){
         require('connection.php');
-        $result = $mysqli->query($selectQuary);
+        $result = $mysqli->query($selectQuery);
         if ($mysqli->errno){
             die('Select Error (' . $mysqli->errno . ') ' . $mysqli->error);
         }
@@ -10,7 +11,19 @@
             $has_rows = true;
             echo '<tr>';
             for($col = 0; $col < count($row); $col++)
-                echo "<td>" . $row[$col] . "</td>";
+                if(!is_null($replaceJoinsCols) && array_key_exists($col, $replaceJoinsCols)){
+                    $joinResult = $mysqli->query($replaceJoinsCols[$col]);
+                    if ($mysqli->errno){
+                        die('Select Error (' . $mysqli->errno . ') ' . $mysqli->error);
+                    }
+                    while($join_row = mysqli_fetch_row($joinResult))
+                        if($join_row[0] == $row[$col]){
+                            echo '<td data-id="' . $row[$col] . '">' . $join_row[1] . '</td>';
+                            break;
+                        }
+                }
+                else
+                    echo "<td>" . $row[$col] . "</td>";
             if($addChangeBtn)
                 echo '<td><div class="change-btn" onclick="setChangeInfo(this)">&#10001;</div></td>';
             if($addDeleteBtn)
